完成bank card api

+++ src/main/java/com/g16/healthpay/service/BankCardServiceImpl.java
@@ -1,18 +1,14 @@
 package com.g16.healthpay.service;
 
 
-
-import java.util.List;
-
-import com.g16.healthpay.dto.BankCardMessage;
-import com.g16.healthpay.mapper.BankCardApiDao;
 import com.g16.healthpay.mapper.BankCardDao;
 import com.g16.healthpay.model.BankCard;
-import com.g16.healthpay.model.BankCardApi;
+import com.g16.healthpay.api.BankCardApi;
 import com.g16.healthpay.utils.EncrypteUtils;
 import com.g16.healthpay.utils.RedisUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
 import javax.annotation.Resource;
 import java.util.List;
 
@@ -22,7 +18,7 @@ public class BankCardServiceImpl implements BankCardService{
     @Resource
     BankCardDao bankCardDao;
     @Resource
-    BankCardApiDao bankCardApiDao;
+    BankCardApi bankCardApi;
     @Autowired
     EncrypteUtils encrypteUtils;
     @Autowired
@@ -30,16 +26,15 @@ public class BankCardServiceImpl implements BankCardService{
 
     @Override
     public boolean bindCard(String phone,String cardNumber,String password){
-        BankCardApi result1;
-        int result2;
-        result1 = bankCardApiDao.checkPassword(cardNumber,password);
-        if(result1!=null){
+        int result;
+        //调用绑卡API
+        if(bankCardApi.bindCard(cardNumber,password)){
             BankCard bankCard = new BankCard();
             bankCard.setCardNumber(cardNumber);
             bankCard.setPhone(phone);
             bankCardDao.selectByPrimaryKey(cardNumber);
-            result2 = bankCardDao.insert(bankCard);
-            if(result2 !=0){
+            result = bankCardDao.insert(bankCard);
+            if(result !=0){
                 return true;
             }
         }
@@ -49,14 +44,14 @@ public class BankCardServiceImpl implements BankCardService{
     @Override
     public boolean deleteCard(String cardNumber){
         int result;
-        result = bankCardDao.deleteByPrimaryKey(cardNumber);
-        if(result!=0){
-            return true;
+        //调用解绑API检查该卡号是否存在
+        if(bankCardApi.unbindCard(cardNumber)){
+            result = bankCardDao.deleteByPrimaryKey(cardNumber);
+            if(result!=0){
+                return true;
+            }
         }
-        else{
-            return false;
-        }
-
+        return false;
     }
 
 
@@ -69,6 +64,4 @@ public class BankCardServiceImpl implements BankCardService{
             return null;
         }
     }
-
-
 }
